+++ backend/src/main/resources/application-auth.yml
@@ -1,24 +1,23 @@
 spring:
   auth:
-    # TODO: public한 API 경로가 더 있는지 검토 필요
-    whitelist-api-urls:
-      - "/auth/login/**"
-      - "/swagger-ui/**"
-      - "/v3/api-docs/**"
-
-    sign-up-url: "/user/signup"
-
+    url:
+      # TODO: public한 API 경로가 더 있는지 검토 필요
+      sign-up: "/user/signup"
+      whitelists:
+        - "/auth/login/**"
+        - "/swagger-ui/**"
+        - "/v3/api-docs/**"
     jwt:
       secret: "${JWT_SECRET}"
       token-expiration:
-        access-token: 1h
-        refresh-token: 7d
-
-    google:
-      client-id: "${GOOGLE_CLIENT_ID}"
-      client-secret: "${GOOGLE_CLIENT_SECRET}"
-      redirect-uri: "${GOOGLE_REDIRECT_URI}"
-      scope:
-        - openid
-        - profile
-        - email
+        access-token: ${JWT_ACCESS_TOKEN_EXPIRATION}
+        refresh-token: ${JWT_REFRESH_TOKEN_EXPIRATION}
+    oauth2:
+      google:
+        client-id: "${GOOGLE_CLIENT_ID}"
+        client-secret: "${GOOGLE_CLIENT_SECRET}"
+        redirect-uri: "${GOOGLE_REDIRECT_URI}"
+        scope:
+          - openid
+          - profile
+          - email
